requirements.txt
# LangChain RAG System Requirements
# 基于LangChain的RAG系统依赖包

# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# HTTP requests for OpenRouter
requests>=2.28.0

# Environment management
python-dotenv>=1.0.0

# Cross-encoder for reranking
transformers>=4.30.0
torch>=2.0.0

# RAGAS evaluation metrics
ragas>=0.1.0

# Optional: For better performance
# faiss-gpu>=1.7.4  # Uncomment if you have CUDA support

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0
